refactor(host-application): split task delete and reset lifecycle services
