+++ apps/desktop/src-tauri/crates/host-application/src/app_service/task_workflow/mod.rs
@@ -1,8 +1,11 @@
 mod approval_service;
 mod document_service;
+mod implementation_reset_service;
+mod lifecycle_support;
 mod qa_service;
 mod session_service;
 mod task_context;
+mod task_deletion_service;
 mod task_service;
 
-pub(crate) use task_service::{normalize_path_for_comparison, normalize_path_key};
+pub(crate) use lifecycle_support::{normalize_path_for_comparison, normalize_path_key};
